+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/condition/ConditionalOnReflectRefreshType.java
@@ -34,7 +34,7 @@ import org.springframework.context.annotation.Conditional;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ ElementType.TYPE, ElementType.METHOD })
 @Documented
-@Conditional(NonDefaultBehaviorCondition.class)
-public @interface ConditionalOnNonDefaultBehavior {
+@Conditional(ReflectRefreshTypeCondition.class)
+public @interface ConditionalOnReflectRefreshType {
 
 }